create tecnico

+++ src/app/services/auth.ts
@@ -1,9 +1,9 @@
 import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { Credenciais } from '../models/credenciais';
 import { API_CONFIG } from '../config/api.config';
 import { JwtHelperService } from '@auth0/angular-jwt';
-
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'
@@ -13,7 +13,9 @@ export class Auth {
   jwtService: JwtHelperService = new JwtHelperService;
   private isBrowser: boolean;
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient, @Inject(PLATFORM_ID) private platformId: object) {
+    this.isBrowser = isPlatformBrowser(platformId);
+  }
 
   authenticate(creds: Credenciais) {
     return this.http.post(`${API_CONFIG.baseUrl}/login`, creds, {
@@ -23,19 +25,26 @@ export class Auth {
   }
 
   successfulLogin(authToken: string) {
-    localStorage.setItem('token', authToken);
+    if (this.isBrowser) {
+      localStorage.setItem('token', authToken);
+    }
   }
 
   isAuthenticated() {
-    let token = localStorage.getItem('token')
-    if (token != null) {
-      return !this.jwtService.isTokenExpired(token)
+    if (this.isBrowser) {
+      let token = localStorage.getItem('token')
+      if (token != null) {
+        return !this.jwtService.isTokenExpired(token)
+      }
     }
-    return false
+    return false;
+
   }
 
   logout() {
-    localStorage.clear();
+    if (this.isBrowser) {
+      localStorage.clear();
+    }
   }
 
 }
